refactor function to make it easier to test
